Slim_Python_Script
Added the scripts from the first version of the model.
The SLiM script runs a simulation of 10 generations with a population size of 200 individuals. It takes two samples of 20 individuals in the first generation and the last one. The simulation is based on a Non Wright Fisher Model but with Wright Fisher caracteristics (one-year generation, constant population size, only one population etc...)
The python script can run the SLiM script, filter from the samples made during the simulation, recapitate from them, and add mutations afterward.