New scripts
Added 'main.py' script that runs n simulations
Updated 'run_simulation.py' so that it runs a simulation then recapitate and add mutations and then calculate the stats. Finally it creates a dataframe with all the useful informations.
Updated the SLiM Script so that it can be used by the python script.